--- grafana-plugin/src/containers/AlertRules/AlertRules.module.css
@@ -1,144 +1,0 @@
-.root {
-  display: flex;
-  gap: 24px;
-  flex-direction: column;
-}
-
-.route {
-  margin-bottom: 10px;
-}
-
-.route-content {
-  margin-left: 16px;
-}
-
-.route-header {
-  margin-bottom: 10px;
-}
-
-.verbal-name {
-  font-weight: 500;
-}
-
-.headerBlock {
-  background-color: var(--secondary-background);
-}
-
-.title {
-  font-size: 16px;
-}
-
-.root .timeline {
-  margin-left: 20px;
-}
-
-.root .steps {
-  margin: 10px 0 0 0;
-}
-
-.root .select {
-  flex-shrink: 0;
-  min-width: 100px;
-}
-
-.create-channel-filter {
-  margin-left: 40px;
-  margin-top: 10px;
-}
-
-.channel-filter-header {
-  display: flex;
-  flex-wrap: wrap-reverse;
-}
-
-.channel-filter-header-left,
-.channel-filter-header-right {
-  flex-grow: 1;
-  margin-bottom: 16px;
-}
-
-.channel-filter-header-right {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.channel-filter-header-title {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  column-gap: 4px;
-  row-gap: 8px;
-}
-
-.root .channel mark {
-  padding: 2px 5px;
-}
-
-.channel_disabled {
-  text-decoration: line-through;
-}
-
-.alertRulesActions {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 24px;
-}
-
-.alertRulesContent {
-  flex-shrink: 0;
-  overflow: auto;
-  max-height: 65vh;
-}
-
-.slackLink {
-  font-size: 12px;
-}
-
-.buttons {
-  display: flex;
-  align-items: baseline;
-}
-
-.buttons > * {
-  margin-left: 8px;
-}
-
-.slack-channel-step {
-  border-radius: 2px;
-  background: var(--secondary-background);
-  padding: 2px 2px 2px 12px;
-  flex-wrap: wrap;
-}
-
-.icons-container {
-  align-self: center;
-}
-
-.slack-channel-switch {
-  margin-left: -8px;
-}
-
-.description-style a {
-  color: var(--primary-text-link);
-}
-
-.integration__heading-text {
-  display: flex;
-  gap: 8px;
-}
-
-.integration__heading-container {
-  display: flex;
-  flex-wrap: wrap-reverse;
-}
-
-.integration__heading-container-left,
-.integration__heading-container-right {
-  flex-grow: 1;
-  margin-bottom: 12px;
-}
-
-.integration__heading-container-right {
-  display: flex;
-  justify-content: flex-end;
-}
