--- grafana-plugin/src/containers/IntegrationSettings/IntegrationSettings.module.css
@@ -1,32 +1,0 @@
-.root {
-  display: block;
-}
-
-.title {
-  padding: 24px;
-  display: flex;
-  gap: 10px;
-  align-items: flex-start;
-}
-
-.title-column {
-  display: flex;
-  flex-direction: column;
-}
-
-.content {
-  margin-top: 10px;
-}
-
-.content ul {
-  list-style-type: none;
-}
-
-.content ol {
-  list-style-type: none;
-}
-
-.settings-header-buttons {
-  right: 0;
-  position: absolute;
-}
