--- grafana-plugin/src/containers/IntegrationSettings/IntegrationSettings.types.ts
@@ -1,7 +1,0 @@
-export enum IntegrationSettingsTab {
-  Templates = 'Templates',
-  Heartbeat = 'Heartbeat',
-  Autoresolve = 'Autoresolve',
-  LiveLogs = 'LiveLogs',
-  HowToConnect = 'HowToConnect',
-}
