--- grafana-plugin/src/containers/IntegrationSettings/parts/Autoresolve.module.css
@@ -1,47 +1,0 @@
-.border-container {
-  margin-bottom: 32px;
-  padding-bottom: 32px;
-  border-bottom: var(--border);
-}
-
-.root ul {
-  margin-left: 24px;
-}
-
-.settings-label {
-  display: grid;
-}
-
-.team-select {
-  width: 520px;
-}
-
-.team-select-actionbuttons {
-  margin-top: 24px;
-}
-
-.confirmation-buttons {
-  display: flex;
-  flex-direction: row-reverse;
-}
-
-.confirmation-buttons .save-team-button {
-  margin-left: 8px;
-}
-
-.autoresolve-block {
-  height: 32px;
-  padding: 4px 8px;
-  margin-top: 8px;
-  min-width: 500px;
-  width: 620px;
-}
-
-.autoresolve-div {
-  display: flex;
-  align-items: baseline;
-}
-
-.warning-icon-color {
-  color: var(--warning-text-color);
-}
