+++ grafana-plugin/src/pages/integration/CommonIntegration.helper.ts
