+++ grafana-plugin/src/pages/integration/Integration.config.ts
@@ -3,7 +3,7 @@
   Any change to this file needs to be done in the oncall-private also
 */
 
-import { BASE_INTEGRATION_TEMPLATES_LIST, BaseTemplateOptions } from './Integration2Common.config';
+import { BASE_INTEGRATION_TEMPLATES_LIST, BaseTemplateOptions } from './IntegrationCommon.config';
 
 export const TemplateOptions = {
   ...BaseTemplateOptions,
