+++ grafana-plugin/src/pages/integration/IntegrationCommon.config.ts
