--- grafana-plugin/src/pages/integrations/Integrations.module.css
@@ -1,38 +1,0 @@
-.filters {
-  margin-bottom: 20px;
-}
-
-.integrations {
-  width: 100%;
-  min-width: 720px;
-  display: flex;
-  align-items: flex-start;
-  border: var(--border);
-  border-radius: 2px;
-}
-
-.no-integrations {
-  width: 100%;
-  padding: 50px 0;
-}
-
-.alert-receive-channels-list {
-  width: 300px;
-  flex-shrink: 0;
-  overflow: auto;
-  max-height: 70vh;
-}
-
-.alert-rules {
-  margin: 16px;
-  flex-grow: 1;
-}
-
-.newIntegrationButton {
-  margin: 16px;
-  width: calc(100% - 32px);
-}
-
-.integrationsList {
-  border-right: var(--border);
-}
