+++ grafana-plugin/src/pages/integrations/Integrations.module.scss
