+++ grafana-plugin/src/pages/routes.tsx
@@ -1,7 +1,6 @@
 import EscalationsChainsPage from 'pages/escalation-chains/EscalationChains';
 import IncidentPage from 'pages/incident/Incident';
 import IncidentsPage from 'pages/incidents/Incidents';
-// import IntegrationsPage from 'pages/integrations/Integrations';
 import MaintenancePage from 'pages/maintenance/Maintenance';
 import OrganizationLogPage from 'pages/organization-logs/OrganizationLog';
 import OutgoingWebhooks from 'pages/outgoing_webhooks/OutgoingWebhooks';
@@ -14,7 +13,7 @@ import CloudPage from 'pages/settings/tabs/Cloud/CloudPage';
 import LiveSettingsPage from 'pages/settings/tabs/LiveSettings/LiveSettingsPage';
 import UsersPage from 'pages/users/Users';
 
-import IntegrationsPage2 from './integrations_2/Integrations2';
+import IntegrationsPage2 from './integrations/Integrations';
 
 export interface NavRoute {
   id: string;
@@ -34,13 +33,9 @@ export const routes: { [id: string]: NavRoute } = [
     component: UsersPage,
     id: 'users',
   },
-  // {
-  //   component: IntegrationsPage,
-  //   id: 'integrations',
-  // },
   {
     component: IntegrationsPage2,
-    id: 'integrations_2',
+    id: 'integrations',
   },
   {
     component: EscalationsChainsPage,
