+++ grafana-plugin/src/plugin/GrafanaPluginRootPage.tsx
@@ -24,9 +24,8 @@ import NoMatch from 'pages/NoMatch';
 import EscalationChains from 'pages/escalation-chains/EscalationChains';
 import Incident from 'pages/incident/Incident';
 import Incidents from 'pages/incidents/Incidents';
-import Integration2 from 'pages/integration_2/Integration2';
-// import Integrations from 'pages/integrations/Integrations';
-import Integrations2 from 'pages/integrations_2/Integrations2';
+import Integration from 'pages/integration/Integration';
+import Integrations from 'pages/integrations/Integrations';
 import Maintenance from 'pages/maintenance/Maintenance';
 import OrganizationLogPage from 'pages/organization-logs/OrganizationLog';
 import OutgoingWebhooks from 'pages/outgoing_webhooks/OutgoingWebhooks';
@@ -150,14 +149,11 @@ export const Root = observer((props: AppRootProps) => {
             <Route path={getRoutesForPage('users')} exact>
               <Users query={query} />
             </Route>
-            {/* <Route path={getRoutesForPage('integrations')} exact>
-              <Integrations query={query} />
-            </Route> */}
             <Route path={getRoutesForPage('integrations')} exact>
-              <Integrations2 query={query} />
+              <Integrations query={query} />
             </Route>
             <Route path={getRoutesForPage('integration')} exact>
-              <Integration2 query={query} />
+              <Integration query={query} />
             </Route>
             <Route path={getRoutesForPage('escalations')} exact>
               <EscalationChains query={query} />
